avoir rerender of sidebar (move up in wrapper)
