+++ src/App.jsx
@@ -1,11 +1,11 @@
-import MainMap from './components/MainMap'
+import Wrapper from './components/Wrapper'
 import './App.css'
 
 function App() {
   return (
     <div className='App'>
       {/* <h1>Plop</h1> */}
-    <MainMap />
+    <Wrapper />
     </div>
   )
 }
