+++ src/components/MainMap.css
@@ -1,0 +1,4 @@
+.MapWrapper {
+    width: 100vw;
+    height: 100vh;
+}
