+++ src/components/MainMap.jsx
@@ -1,111 +1,26 @@
-import { useEffect, useState } from "react";
-// import axios from 'axios';
+import {  useEffect, useState } from "react";
 import { Map, GeoJson } from "pigeon-maps"
-import SideBar from "./SideBar";
-import { flagToCountryCode } from "../helpers/countryUtil";
-import { getZoom, getCenter, getListBoundingBox, filterGpxList, getGpxList, loadGeoJson } from "../helpers/gpxUtil";
 import Popup from "./Popup";
 import { CartoDBVoyager } from "../helpers/tiles";
+import { getCenter, getListBoundingBox, getZoom } from "../helpers/gpxUtil";
+import './MainMap.css'
 
-const MainMap = () => {
-  const [step, setStep] = useState(0);
-  const [allGpxList, setAllGpxList] = useState([]);
-  const [gpxList, setGpxList] = useState([]);
-  const [geojson, setGeojson] = useState(null);
+const MainMap = ({geojson, geojsonList}) => {
   const [center, setCenter] = useState({lon:3, lat:50});
   const [zoom, setZoom] = useState(8);
-  const [geojsonList, setGeojsonList] = useState([]);
-  const [allGeojsonList, setAllGeojsonList] = useState([]);
-  
-  const [currentYear, setCurrentYear] = useState('2024');
-  const [currentCountry, setCurrentCountry] = useState('xx');
+
   const [currentFocus, setCurrentFocus] = useState(null);
-  // const allGeojsonList = []
-// TODO use swr
 
   useEffect(() => {
-    // do not load anything on initial render
-    if (step===0) {
-      setStep(1)
-      const yearNow = new Date().getFullYear()
-      setCurrentYear(yearNow);
+    console.log("useEffect");
+    if (geojsonList.length) {
+      console.log("useEffect", geojsonList.length);
+      const bbox = getListBoundingBox(geojsonList)
+      setCenter(getCenter(bbox))
+      setZoom(getZoom(bbox))
+      console.log("bbox", bbox);
     }
-  }, [step]);
-
-  useEffect(() => {
-    if (step===1) {
-      const getGpxListAwaited = async () => {
-        const gpxList = await getGpxList()
-        gpxList.length && setAllGpxList(gpxList)
-      }
-      getGpxListAwaited()
-      setStep(3)
-    }
-  }, [step]);
-
-  useEffect(() => {
-    if (step===3) {
-      const setGeoJsonListAwaited = async (gpxList) => {
-        const geojsonList = []
-        const requests = gpxList.map( async (url, i) => {
-          if (i < gpxList.length) {
-            const geojson = allGeojsonList.find(track => track.url === url) || await loadGeoJson(url)
-            if (!allGeojsonList.find(track => track.url === url)) {
-              geojson["url"] = url;
-              allGeojsonList.push(geojson)
-            }
-            geojsonList.push(geojson)
-            setGeojsonList(geojsonList)
-            // need to find a way to render on each track
-          }
-        })
-        Promise.all(requests).then(() => {
-          setGeojsonList(geojsonList)
-          if (geojsonList.length) {
-            const bbox = getListBoundingBox(geojsonList)
-            setCenter(getCenter(bbox))
-            setZoom(getZoom(bbox))
-          }
-          setAllGeojsonList(allGeojsonList)
-          setStep(5) // make sure we are done
-        })
-      }
-
-      setStep(4)
-      setGeoJsonListAwaited(gpxList)
-    }
-  }, [step, gpxList]);
-
-  useEffect(() => {
-    if (step===6) {
-      const gpsList = filterGpxList(currentYear, currentCountry, allGpxList)
-      if (gpsList.length) {
-        setGpxList(gpsList)
-        setStep(3)
-      } else {
-        setGeojsonList([])
-        setGeojson(null)
-        setStep(5)
-      }
-    }
-  }, [step, currentYear, currentCountry, allGpxList]);
-
-  const handleClickSideBar = (e) => {
-    // close popup
-    // e.target.parentNode.parentNode.nextSibling.style.display = 'none'
-    // unset pupup
-    setCurrentFocus(null)
-    // get click info
-    if (e.target.innerText >= 2010) {
-      setCurrentYear(e.target.innerText)
-    } else if (e.target.innerText === 'all') {
-      setCurrentYear('')
-    } else {
-      const cc = flagToCountryCode(e.target.innerText)
-      setCurrentCountry(cc)
-    }
-    setStep(6)
-  }
+  }, [geojsonList]);
 
   const renderGeoJson = (geojson, key) => {
     if (!geojson)  return null;
@@ -133,7 +48,7 @@ const MainMap = () => {
     );
   };
 
-  const handleClick = (e, geojson) => {
+  const handleClick = (e) => {
     if (currentFocus) {
       currentFocus.map(el => {
         el.setAttribute('stroke', 'red');
@@ -142,7 +57,7 @@ const MainMap = () => {
     }
     const svgPathArray = [...e.event.target.parentNode.children];
     setCurrentFocus(svgPathArray)
-    setGeojson(geojson)
+    // setGeojson(geojson)
     svgPathArray.map(el => {
       el.setAttribute('stroke', 'blue');
       el.setAttribute('opacity', '1');
@@ -161,8 +76,9 @@ const MainMap = () => {
     // popEl.children[1].children[0].children[3].innerText = geojson.distance+'km' || '';
   }
 
+  console.log("RENDER MAP");
   return (
-    <>
+    <div className="MapWrapper">
       <Map
         provider={CartoDBVoyager.tiles}
         defaultZoom={8}
@@ -171,21 +87,11 @@ const MainMap = () => {
         center={[center.lat || 44, center.lon || 3]}
         zoom={zoom || 8}
       >
-
-        <SideBar 
-          step={step}
-          currentYear={currentYear}
-          currentCountry={currentCountry}
-          geojsonList={geojsonList}
-          handleClick={handleClickSideBar}
-        />
         {currentFocus && <Popup key='popup' currentFocus={currentFocus} geojson={geojson} />}
-        { step>=3 && geojson && renderGeoJson(geojson, 'prems') }
-        { step===4 && geojsonList && geojsonList.length>=1 && geojsonList.map((json, i) => renderGeoJson(json, i)) }
-        { step===5 && geojsonList && geojsonList.map((json, i) => renderGeoJson(json, 'comp'+i)) }
-
+        { geojson && renderGeoJson(geojson, 'prems') }
+        { geojsonList && geojsonList.length>=1 && geojsonList.map((json, i) => renderGeoJson(json, i)) }
       </Map>
-    </>
+    </div>
   );
 }
