+++ src/components/SideBar.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import './SideBar.css';
 import { countryCodeToFlag } from '../helpers/countryUtil';
 import { getDistanceList } from '../helpers/gpxUtil';
+import './SideBar.css';
 
 const SideBar = ({step, currentYear, currentCountry, geojsonList, handleClick}) => { 
     const [width, setWidth] = useState('none');
@@ -18,6 +18,7 @@ const SideBar = ({step, currentYear, currentCountry, geojsonList, handleClick}) 
     }   
     const countries = ['nl', 'be', 'fr', 'de', 'ch', 'at', 'cz', 'pl', 'sk', 'hu', 'it', 'lu', 'si']
 
+    console.log("RENDER sidebar");
     return (
         <>
         <div className={`bubble ${width}`} onClick={toggleSideBar}>
